requirements.txt
flask~=2.0.2
flask-pymongo
flask_cors
pymongo~=4.0.1
PyJWT==1.7.1

Werkzeug~=2.0.3